Improving custom Context and Logger, and  (#11)
The biggest thing in this PR is the change from Datadog (proprietary) to
Uptrace (open) as the telemetry engine built into the chassis. Uptrace
is actually just OpenTelemetry so if I want I can switch it out entirely
to be agnostic. We'll see if the telemetry engine gets it's own chassis
interface one day.

Included in this is a full rework to the Context in `chassis`. It used
to be a very clumsy `ExecutionContext` which required a transaction ID
and was passed by pointer (ugh). Now it follows a simple interface that
satisfies the regular `context.Context` of the standard library and
wraps up all the telemetry bits for logging and traces at the same time.
Nice!